requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1

# Scraping
playwright==1.44.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.2.1

# LLM / LangChain
langchain==0.1.20
langchain-anthropic==0.1.8
langchain-openai==0.1.7
anthropic==0.26.0
openai==1.30.1

# Async
anyio==4.3.0

# Export
pandas==2.2.2

# Queue (Celery-ready, optional for Phase 2)
celery==5.3.6
redis==5.0.4

# Database (optional for Phase 2)
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Observability
sentry-sdk==2.3.1
